Fixing build error on Win64.
Review URL: http://codereview.chromium.org/8508058

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@9958 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -1940,7 +1940,7 @@ intptr_t FreeList::EvictFreeListItems(Page* p) {
         EvictFreeListItemsInList(&large_list_, p);
   }
 
-  available_ -= sum;
+  available_ -= static_cast<int>(sum);
 
   return sum;
 }
